# Speed up Fitzpatrick99, like, a lot