--- a/extern/README.md
+++ b/extern/README.md
@@ -1,0 +1,3 @@
+# external code
+
+Fetched using `fetch.sh` from https://github.com/kbarbary/bsplines