… extern/bs.c