This is pretty sweet:

**v0.2.2:**

```
Fitzpatrick99 init                                          472.64us
Fitzpatrick99 call  wave=[ 1000.0,  30000.0]  size=   10     49.46us
Fitzpatrick99 call  wave=[ 1000.0,  30000.0]  size=  100     52.25us
Fitzpatrick99 call  wave=[ 1000.0,  30000.0]  size= 1000    108.72us
Fitzpatrick99 call  wave=[ 1000.0,  30000.0]  size=10000    708.58us
```

**master:**
```
Fitzpatrick99 init                                           17.65us
Fitzpatrick99 call  wave=[ 1000.0,  30000.0]  size=   10      9.73us
Fitzpatrick99 call  wave=[ 1000.0,  30000.0]  size=  100     16.34us
Fitzpatrick99 call  wave=[ 1000.0,  30000.0]  size= 1000     83.65us
Fitzpatrick99 call  wave=[ 1000.0,  30000.0]  size=10000    745.01us
```

**This branch:**
```
Fitzpatrick99 init                                            0.93us
Fitzpatrick99 call  wave=[ 1000.0,  30000.0]  size=   10      5.41us
Fitzpatrick99 call  wave=[ 1000.0,  30000.0]  size=  100      6.73us
Fitzpatrick99 call  wave=[ 1000.0,  30000.0]  size= 1000     23.95us
Fitzpatrick99 call  wave=[ 1000.0,  30000.0]  size=10000    191.28us
```

The spline initizialization is now so fast that we don't really need the `Fitzpatrick99` class. We could just initialize and deallocate the spline on every call.